requirements.txt
torch==1.10.0
torch-geometric==2.1.3
numpy==1.21.5
matplotlib==3.5.1
pandas==1.3.5
tqdm==4.62.3
networkx==2.6.3
biopython==1.79
sklearn==0.24.2
torch_optimizer==0.1.1